Add experimental redline support
Attempt experimental redline support, using intermediate LaTeX files
generated from the Markdown, diffing those, and then attempting to
convert the resulting diff'd file into a PDF.

This does have some real sharp edges, and generates a ton of warnings
running XeTeX on the resulting diff, but helps provide a bit of a
rough estimate. For Ballots, it's likely that using a more formal
tool like Acrobat would be cleaner, using the PDFs to diff, or using
Word to compare the Word docs.

Tools like pandiff use the HTML output, but HTML output is currently
disabled.

In doing so, switch to using Action output files with explicit
filenames, just to allow easier migration of paths in the future.