Fix some review nits & add another pass for redlines
This addresses some nits that came up during review. In
the process of testing these fixes, I tried to stress
test the command by doing a redline between the (unmodified)
BRs and the (unmodified) EVGs. While it was a mess, which is
expected, it also revealed that a third xelatex pass was
necessary in order to ensure both hyperref and the table of
contents work correctly. The first pass builds the ToC file
and links, the second pass then inserts the ToC and links,
and the third pass corrects the page numbers in the ToC to
adjust for the fact that the ToC is now present.

This matches the Pandoc PDF writer (e.g. see
https://github.com/jgm/pandoc/blob/8b872301eb68f0617c648cd6929ca8ff0ecbffd6/src/Text/Pandoc/PDF.hs#L242
).