Update Action (#2)
The action is currently failing because it's built on demand (i.e. not published to Docker Hub or GitHub Packages), and so it is built on demand.

The current Pandoc base image being used by the 2.0.0-rc1 has an older tlmgr, and as a result, the initial image building step fails due to needing to update tlmgr.

This adds a step to explicitly update tlmgr as part of building the image. However, it also updates the base Pandoc image further than 2.0-RC (which also/alternatively resolves this issue), in order to take advantage of the Latex writer now supporting multi-column/multi-row tables.

In order to make sure such changes are safe in the future, it also introduces a base set of unit tests that build a PDF, a DOCX, a PDF redline, and lints for broken links, testing the core functionality of the action, and wires them up to GitHub Actions.