Add --no-highlight to the Pandoc arguments (#14)
* Add --no-highlight to the Pandoc arguments

Pandoc supports syntax highlighting for code blocks that
explicitly specify a language. When used, it injects a
series of LaTeX macros into the template, which it then
uses as part of the conversion.

These macros weren't included when originally building
the template, meaning that attempting to specify a
language would cause a PDF conversion issue, due to
undefined LaTeX macros. While these can be added, the
color styles require further customization to fit within
the document look-and-feel. These can be customized with
JSON files using --highlight-style and, if necessary,
--syntax-definition for KDE XML syntax definition files.

Rather than work to customize the style and syntax
appropriate for the examples, for now, just disable
syntax highlighting. This can be revisited later, with
thematically-appropriate colors and robust language
support.

Fixes #7