Add support for publishing to GitHub Container Registry (#22)
* Add publish step using GITHUB_TOKEN

According to
https://docs.github.com/en/packages/guides/using-github-packages-with-github-actions#upgrading-a-workflow-that-accesses-ghcrio
it is now possible to use GITHUB_TOKEN to publish packages, do so.

Note: This change changes how to manually invoke the docker
image (e.g. when running locally), as follows:

Previously, we relied on the action.yml to declare that the
input_markdown_file argument should be passed as the first positional
argument to the Docker ENTRYPOINT. However, when using the docker://
syntax for a GitHub action (e.g. hosted on ghcr.io), GitHub no longer
knows about the action.yml metadata, and so only invokes with the
arguments supplied (via INPUT_xxx environment variables).

This updates the entrypoint to use the INPUT_MARKDOWN_FILE instead of
the first argument to the entrypoint script, and to no longer pass
it as a positional argument when invoking using the "reference to a
repository action".

The goal here is to avoid needing to update workflows to add an
explicit "args" command to go with the docker:// invocation, and
instead can use the same syntax for both docker:// and repository
syntax.